Delete MOCMA and selection implementations in Python
